Morley.Nettest

Description

Michelson contracts testing on a real Tezos network.

It defines an interface for writing network testing scenarios and provides two implementations: one of them talks to reference Tezos software and submits operations to real network, the other one converts scenario to the existing integrational testing eDSL.

Expected usage is the following: 1. Write a scenario using abstract nettest interface. 2. Make an executable (i. e. something that can be executed: executable or test-suite in .cabal file, normally the latter) which runs this scenario via pure integrational testing engine first (as a quick check for correctness) and then via Client implementation. This executable is supposed to run periodically, but not in each MR on each change. 3. Also run this scenario in existing test-suite using pure implementation. So pure implementation will be used twice: in normal test-suite that runs on each change (which proves that it is likely correct) and in real network test to prevent that test from starting if the scenario is most likely not correct. 4. Note that in order to run it on a real network you should have an address with nettest alias and it should have sufficient balance to run the scenario.

TODO: ★ [#50] Collect gas and other statistics in nettest. ★ [#52] Use RPC directly except for signing and key management, use features such as batching. ★ [#53] Maybe somehow merge with testing eDSL. ★ [#54] tezos-client internals should be wrapped into some API and moved somewhere into reusable place. ★ [#55] Add command line options parsing.

Documentation

data AddrOrAlias

Representation of an address that tezos-client uses. It can be an address itself or a textual alias.

Constructors

AddrResolved Address
Address itself, can be used as is.

AddrAlias Text
Address alias, should be resolved by tezos-client.

data OriginateData

Constructors

(NiceParameterFull param, NiceStorage st) => OriginateData

Fields

odFrom :: AddrOrAlias

The address from which contract will be originated (must have some XTZ).

odName :: Text

Alias for the originated contract.

odBalance :: Word64

Initial balance.

odStorage :: st

Initial storage.

odContract :: ContractCode param st

The contract itself.

We are using Lorentz version here which is convenient. However, keep in mind that:

If someone wants to test a contract from `.tz` file, they will have to convert it to Lorentz somehow, which sounds bad, so in this case we'll have to provide another version that takes typed contract. Let's do it only if the need arises.

In TM-366 we want to move optimizer call and other preprocessing to Lorentz compilation. So conversion to typed contract will require an additional argument. Note that we'll have the same problem in integrational testing eDSL as well, so let's handle them both at once. I think that usually we'll want to apply the same compilation options to all contracts, so we can simply add them to some runner. Or we can use default options and add a command that modifies them for further calls.

data TransferData

Constructors

(Show v, NiceParameter v) => TransferData

Fields

tdFrom :: AddrOrAlias

Sender address for this transaction.

tdTo :: AddrOrAlias

Receiver address for this transaction.

tdAmount :: Word64

Amount to be transferred.

tdEntrypoint :: EpName

An entrypoint to be called. Consider using ep in testing scenarios.

tdParameter :: v

Parameter that will be used for a contract call. Set to () for transfers to key addresses.

data NettestImpl m

A record data type with all base methods one can use during nettest.

Constructors

NettestImpl

Fields

niResolveAddr :: AddrOrAlias -> m Address

Unwrap an Address or get an Address corresponding to an alias.

niGenKey :: Text -> m Address

Generate a secret key and store it with given alias. If a key with this alias already exists, the corresponding address will be returned and no state will be changed.

niGenFreshKey :: Text -> m Address

Generate a secret key and store it with given alias. Unlike niGenKey this function overwrites the existing key when given alias is already stored.

niOriginate :: OriginateData -> m Address

Originate a new contract with given data.

niTransfer :: TransferData -> m ()

Send a transaction with given data.

niComment :: Text -> m ()

Print given string verbatim as a comment.

niExpectFailure :: forall a. m a -> NettestFailure -> m ()

Expect a failure while running another action

type NettestScenario = forall m. Monad m => NettestImpl m -> m ()

Any monadic computation that can only access nettest methods.

Constant address

nettestAddr :: AddrOrAlias

This address will be used as source by default. Some functionality in this module assumes that this address has sufficient amount of XTZ in advance (required amount depends on scenario).

Actions

resolveAddrAction :: NettestImpl m -> AddrOrAlias -> m Address

Alias for niResolveAddr.

resolveNettestAddrAction :: NettestImpl m -> m Address

Get Address corresponding to nettestAddr.

newAddressAction :: Monad m => NettestImpl m -> Text -> m Address

Generate a new secret key and record it with given alias. Also transfer small amount of XTZ to it from nettestAddr. Does not override an existing address.

originateAction :: NettestImpl m -> OriginateData -> m Address

Alias for niOriginate.

originateSimpleAction :: (NiceParameterFull param, NiceStorage st) => NettestImpl m -> Text -> st -> ContractCode param st -> m Address

A simplified version of the originate command. Uses default address (nettestAddr). The contract will have 0 balance.

transferAction :: NettestImpl m -> TransferData -> m ()

Alias for niTransfer.

callAction :: (Show v, NiceParameter v) => NettestImpl m -> AddrOrAlias -> EpName -> v -> m ()

Call a certain entrypoint of a contract referred to by some address. The sender is nettestAddr, so it should have sufficient XTZ to pay tx fee.

callFromAction :: (Show v, NiceParameter v) => NettestImpl m -> AddrOrAlias -> AddrOrAlias -> EpName -> v -> m ()

Call a certain entrypoint of a contract referred to by the second address from the first address.

commentAction :: NettestImpl m -> Text -> m ()

Alias for niComment.

expectFailureAction :: NettestImpl m -> m a -> NettestFailure -> m ()

Alias for niExpectFailure.

Helpers

ep :: HasCallStack => Text -> EpName

A short partial constructor for EpName. It is supposed to be applied to string constants, so programmer is responsible for validity. And this code is for tests anyway, so each failure is a programmer mistake.

It is intentionally here and not in some deeper module because the name is really short and more suitable for writing scenarios.

Validation

data NettestFailure

Failures that could be expected in the execution of a NettestScenario action

Constructors

NettestFailedWith

NettestMutezArithError ArithErrorType

NettestShiftArithError

NettestGasExhaustion

Real network implementation based on tezos-client

data NettestClientConfig

Configuration of nettest implementation for real network.

Constructors

NettestClientConfig

Fields

nccScenarioName :: Maybe Text

nccNodeAddress :: Maybe Text

nccNodePort :: Maybe Word16

nccTezosClient :: FilePath

nccNodeUseHttps :: Bool

nccVerbose :: Bool

runNettestClient :: NettestClientConfig -> NettestScenario -> IO ()

Run NettestScenario using implementation for real network. Throws an error if an unexpected failure occurs or if an expected one doesn't

nettestImplClient :: NettestClientConfig -> NettestImpl IO

Implementation that works with real network and uses `tezos-node` RPC and `tezos-client`.

caps-based commands

type MonadNettest caps base m = (Monad base, m ~ CapsT caps base, HasCap NettestImpl caps)

Constraint for a monad in which we can do nettest actions. It requires NettestImpl capability.

type NettestT m = CapsT '[NettestImpl] m

Monad transformer that adds only NettestImpl capability.

resolveAddr :: MonadNettest caps base m => AddrOrAlias -> m Address

resolveAddrAction adapted to caps.

resolveNettestAddr :: MonadNettest caps base m => m Address

resolveAddrAction adapted to caps.

newAddress :: MonadNettest caps base m => Text -> m Address

newAddressAction adapted to caps.

originate :: MonadNettest caps base m => OriginateData -> m Address

originateAction adapted to caps.

originateSimple :: (MonadNettest caps base m, NiceParameterFull param, NiceStorage st) => Text -> st -> ContractCode param st -> m Address

originateSimpleAction adapted to caps.

transfer :: MonadNettest caps base m => TransferData -> m ()

transferAction adapted to caps.

call :: (MonadNettest caps base m, NiceParameter v, Show v) => AddrOrAlias -> EpName -> v -> m ()

callAction adapted to caps.

callFrom :: (MonadNettest caps base m, NiceParameter v, Show v) => AddrOrAlias -> AddrOrAlias -> EpName -> v -> m ()

callFromAction adapted to caps.

comment :: MonadNettest caps base m => Text -> m ()

commentAction adapted to caps.

expectFailure :: MonadNettest caps base m => m a -> NettestFailure -> m ()

expectFailureAction adapted to caps.

uncapsNettest :: forall m a. Monad m => NettestT m a -> NettestImpl m -> m a

Adapt caps-based interface back to argument passing style.

Integration with integrational testing engine (pun intended)

runNettestViaIntegrational :: NettestScenario -> IO ()

Run a NettestScenario by converting it to integrational scenario and then passing to HUnit engine. Throws an exception in case of an unexpected failure.

It is suitable for very simple test-suites where we just want to test scenario quickly first and then run it on real network in case we consider it valid.

nettestToIntegrational :: NettestScenario -> IntegrationalScenario

Build IntegrationalScenario out of NettestScenario. Also accepts a NettestExpectation to validate the result of the scenario

CLI

getClientConfig :: Maybe Text -> IO NettestClientConfig

clientConfigParser :: Maybe Text -> Parser NettestClientConfig

Standard command line options parser for NettestClientConfig.