Morley.Nettest.Abstract

Description

Abstract nettest interface not bound to a particular implementation.

The interface may look a bit untyped and unsafe in some places. For example, in order to call a contract one should supply a simple address rather than a contract ref, so it is easy to pass a value of wrong type. Also it is easy to call a non-existing entrypoint.

Subjectively, it makes writing test scenarios easier because you have to prove less to the compiler. It also makes implementation of nettest engine a bit easier. Of course, it also makes it easier to make certain mistakes. However, we expect users of this interface to also use the functionality of the Pure module and convert nettest scenarios to purely testable scenarios for integrational testing engine. In this case errors should be detected almost as quickly as they would reported by the compiler, at least before trying to run scenario on a live network.

Also this interface uses Address rather than EpAddress. I (@gromak) concluded that EpAddress can not be passed to tezos-client. For key addresses it just does not make sense and for contract addresses I get such errors:

  bad contract notation
  Invalid contract notation "KT1VrFpBPwBTm3hsK7DB7SPmY8fTHJ3vY6sJ%mint"

Documentation

data AddrOrAlias

Representation of an address that tezos-client uses. It can be an address itself or a textual alias.

Constructors

AddrResolved Address
Address itself, can be used as is.

AddrAlias Text
Address alias, should be resolved by tezos-client.

data OriginateData

Constructors

(NiceParameterFull param, NiceStorage st) => OriginateData

Fields

odFrom :: AddrOrAlias

The address from which contract will be originated (must have some XTZ).

odName :: Text

Alias for the originated contract.

odBalance :: Word64

Initial balance.

odStorage :: st

Initial storage.

odContract :: ContractCode param st

The contract itself.

We are using Lorentz version here which is convenient. However, keep in mind that:

If someone wants to test a contract from `.tz` file, they will have to convert it to Lorentz somehow, which sounds bad, so in this case we'll have to provide another version that takes typed contract. Let's do it only if the need arises.

In TM-366 we want to move optimizer call and other preprocessing to Lorentz compilation. So conversion to typed contract will require an additional argument. Note that we'll have the same problem in integrational testing eDSL as well, so let's handle them both at once. I think that usually we'll want to apply the same compilation options to all contracts, so we can simply add them to some runner. Or we can use default options and add a command that modifies them for further calls.

data TransferData

Constructors

(Show v, NiceParameter v) => TransferData

Fields

tdFrom :: AddrOrAlias

Sender address for this transaction.

tdTo :: AddrOrAlias

Receiver address for this transaction.

tdAmount :: Word64

Amount to be transferred.

tdEntrypoint :: EpName

An entrypoint to be called. Consider using ep in testing scenarios.

tdParameter :: v

Parameter that will be used for a contract call. Set to () for transfers to key addresses.

data NettestImpl m

A record data type with all base methods one can use during nettest.

Constructors

NettestImpl

Fields

niResolveAddr :: AddrOrAlias -> m Address

Unwrap an Address or get an Address corresponding to an alias.

niGenKey :: Text -> m Address

Generate a secret key and store it with given alias. If a key with this alias already exists, the corresponding address will be returned and no state will be changed.

niGenFreshKey :: Text -> m Address

Generate a secret key and store it with given alias. Unlike niGenKey this function overwrites the existing key when given alias is already stored.

niOriginate :: OriginateData -> m Address

Originate a new contract with given data.

niTransfer :: TransferData -> m ()

Send a transaction with given data.

niComment :: Text -> m ()

Print given string verbatim as a comment.

niExpectFailure :: forall a. m a -> NettestFailure -> m ()

Expect a failure while running another action

type NettestScenario = forall m. Monad m => NettestImpl m -> m ()

Any monadic computation that can only access nettest methods.

Constant address

nettestAddr :: AddrOrAlias

This address will be used as source by default. Some functionality in this module assumes that this address has sufficient amount of XTZ in advance (required amount depends on scenario).

nettestAddrText :: Text

Text inside nettestAddr.

Actions

resolveAddrAction :: NettestImpl m -> AddrOrAlias -> m Address

Alias for niResolveAddr.

resolveNettestAddrAction :: NettestImpl m -> m Address

Get Address corresponding to nettestAddr.

newAddressAction :: Monad m => NettestImpl m -> Text -> m Address

Generate a new secret key and record it with given alias. Also transfer small amount of XTZ to it from nettestAddr. Does not override an existing address.

newFreshAddressAction :: NettestImpl m -> Text -> m Address

Generate a new secret key and record it with given alias. If the alias is already known, the key will be overwritten. The address is guaranteed to be fresh, i. e. no operations on it have been made.

originateAction :: NettestImpl m -> OriginateData -> m Address

Alias for niOriginate.

originateSimpleAction :: (NiceParameterFull param, NiceStorage st) => NettestImpl m -> Text -> st -> ContractCode param st -> m Address

A simplified version of the originate command. Uses default address (nettestAddr). The contract will have 0 balance.

transferAction :: NettestImpl m -> TransferData -> m ()

Alias for niTransfer.

callAction :: (Show v, NiceParameter v) => NettestImpl m -> AddrOrAlias -> EpName -> v -> m ()

Call a certain entrypoint of a contract referred to by some address. The sender is nettestAddr, so it should have sufficient XTZ to pay tx fee.

callFromAction :: (Show v, NiceParameter v) => NettestImpl m -> AddrOrAlias -> AddrOrAlias -> EpName -> v -> m ()

Call a certain entrypoint of a contract referred to by the second address from the first address.

commentAction :: NettestImpl m -> Text -> m ()

Alias for niComment.

expectFailureAction :: NettestImpl m -> m a -> NettestFailure -> m ()

Alias for niExpectFailure.

Validation

data NettestFailure

Failures that could be expected in the execution of a NettestScenario action

Constructors

NettestFailedWith

NettestMutezArithError ArithErrorType

NettestShiftArithError

NettestGasExhaustion

Helpers

ep :: HasCallStack => Text -> EpName

A short partial constructor for EpName. It is supposed to be applied to string constants, so programmer is responsible for validity. And this code is for tests anyway, so each failure is a programmer mistake.

It is intentionally here and not in some deeper module because the name is really short and more suitable for writing scenarios.