Morley.Nettest.Client

Description

Implementation that works with real Tezos network, it talks to a Tezos node and uses `tezos-client`.

Documentation

data NettestClientConfig

Configuration of nettest implementation for real network.

Constructors

NettestClientConfig

Fields

nccScenarioName :: Maybe Text

nccNodeAddress :: Maybe Text

nccNodePort :: Maybe Word16

nccTezosClient :: FilePath

nccNodeUseHttps :: Bool

nccVerbose :: Bool

data NettestClientError

Constructors

UnexpectedClientFailure Text

UnknownAddress

Fields

Text

Name of address which is eventually used

Text

Original name passed by user

UnexpectedClientSuccess

runNettestClient :: NettestClientConfig -> NettestScenario -> IO ()

Run NettestScenario using implementation for real network. Throws an error if an unexpected failure occurs or if an expected one doesn't

nettestImplClient :: NettestClientConfig -> NettestImpl IO

Implementation that works with real network and uses `tezos-node` RPC and `tezos-client`.