Morley.Nettest.Pure

Description

Integration with integrational testing engine (pun intended).

Documentation

runNettestViaIntegrational :: NettestScenario -> IO ()

Run a NettestScenario by converting it to integrational scenario and then passing to HUnit engine. Throws an exception in case of an unexpected failure.

It is suitable for very simple test-suites where we just want to test scenario quickly first and then run it on real network in case we consider it valid.

nettestToIntegrational :: NettestScenario -> IntegrationalScenario

Build IntegrationalScenario out of NettestScenario. Also accepts a NettestExpectation to validate the result of the scenario